+++ pyodide-build/pyodide_build/cli/build.py
@@ -1,4 +1,5 @@
 import os
+import re
 import shutil
 import tempfile
 from pathlib import Path
@@ -7,38 +8,11 @@ from urllib.parse import urlparse
 
 import requests
 import typer  # type: ignore[import]
-from unearth.evaluator import TargetPython
-from unearth.finder import PackageFinder
 
 from .. import common
 from ..out_of_tree import build
+from ..out_of_tree.pypi import build_dependencies_for_wheel, fetch_pypi_package
 from ..out_of_tree.utils import initialize_pyodide_root
-
-
-def _fetch_pypi_package(package_spec, destdir):
-    PYMAJOR = common.get_make_flag("PYMAJOR")
-    PYMINOR = common.get_make_flag("PYMINOR")
-    tp = TargetPython(
-        py_ver=(int(PYMAJOR), int(PYMINOR)),
-        platforms=[common.platform()],
-        abis=[f"cp{PYMAJOR}{PYMINOR}"],
-    )
-    pf = PackageFinder(index_urls=["https://pypi.org/simple/"], target_python=tp)
-    match = pf.find_best_match(package_spec)
-    if match.best is None:
-        if len(match.candidates) != 0:
-            error = f"""Can't find version matching {package_spec}
-versions found:
-"""
-            for c in match.candidates:
-                error += "  " + str(c.version) + "\t"
-            raise RuntimeError(error)
-        else:
-            raise RuntimeError(f"Can't find package: {package_spec}")
-    with tempfile.TemporaryDirectory() as download_dir:
-        return pf.download_and_unpack(
-            link=match.best.link, location=destdir, download_dir=download_dir
-        )
 
 
 def pypi(
@@ -48,7 +22,7 @@ def pypi(
         help="Which symbols should be exported when linking .so files?",
     ),
     ctx: typer.Context = typer.Context,
-) -> None:
+) -> Path:
     """Fetch a wheel from pypi, or build from source if none available."""
     initialize_pyodide_root()
     common.check_emscripten_version()
@@ -60,19 +34,19 @@ def pypi(
         temppath = Path(tmpdir)
 
         # get package from pypi
-        package_path = _fetch_pypi_package(package, temppath)
+        package_path = fetch_pypi_package(package, temppath)
         if not package_path.is_dir():
             # a pure-python wheel has been downloaded - just copy to dist folder
-            shutil.copy(str(package_path), str(curdir / "dist"))
+            dest_file = curdir / "dist" / package_path.name
+            shutil.copyfile(str(package_path), curdir / "dist" / package_path.name)
             print(f"Successfully fetched: {package_path.name}")
-            return
+            return dest_file
 
         # sdist - needs building
         os.chdir(tmpdir)
-        build.run(exports, backend_flags)
-        for src in (temppath / "dist").iterdir():
-            print(f"Built {str(src.name)}")
-            shutil.copy(str(src), str(curdir / "dist"))
+        built_wheel = build.run(exports, backend_flags, outdir=curdir / "dist")
+        os.chdir(curdir)
+        return built_wheel
 
 
 def url(
@@ -82,14 +56,13 @@ def url(
         help="Which symbols should be exported when linking .so files?",
     ),
     ctx: typer.Context = typer.Context,
-) -> None:
+) -> Path:
     """Fetch a wheel or build sdist from url."""
     initialize_pyodide_root()
     common.check_emscripten_version()
     backend_flags = ctx.args
     curdir = Path.cwd()
     (curdir / "dist").mkdir(exist_ok=True)
-
     with requests.get(package_url, stream=True) as response:
         parsed_url = urlparse(response.url)
         filename = os.path.basename(parsed_url.path)
@@ -98,11 +71,11 @@ def url(
             ext = name_base[name_base.rfind(".") :] + ext
         if ext.lower() == ".whl":
             # just copy wheel into dist and return
-            out_path = f"dist/{filename}"
+            out_path = Path(f"dist/{filename}").resolve()
             with open(out_path, "b") as f:
                 for chunk in response.iter_content(chunk_size=1048576):
                     f.write(chunk)
-            return
+            return out_path
         else:
             tf = tempfile.NamedTemporaryFile(suffix=ext, delete=False)
             for chunk in response.iter_content(chunk_size=1048576):
@@ -117,28 +90,31 @@ def url(
                     os.chdir(folder_list[0])
                 else:
                     # unzipped here
-                    os.chdir(tmpdir)
-                print(os.listdir(tmpdir))
-                build.run(exports, backend_flags)
-                for src in (temppath / "dist").iterdir():
-                    print(f"Built {str(src.name)}")
-                    shutil.copy(str(src), str(curdir / "dist"))
+                    os.chdir(temppath)
+                wheel_path = build.run(exports, backend_flags, outdir=curdir / "dist")
             os.unlink(tf.name)
+            return wheel_path
 
 
 def source(
-    source_location: "Optional[str]" = typer.Argument(None),
+    source_location: str,
     exports: str = typer.Option(
         "requested",
         help="Which symbols should be exported when linking .so files?",
     ),
     ctx: typer.Context = typer.Context,
-) -> None:
+) -> Path:
     """Use pypa/build to build a Python package from source"""
     initialize_pyodide_root()
+    orig_dir = Path.cwd()
+    if source_location != ".":
+        # build in this folder
+        os.chdir(source_location)
     common.check_emscripten_version()
-    backend_flags = [source_location] + ctx.args
-    build.run(exports, backend_flags)
+    backend_flags = ctx.args
+    built_wheel = build.run(exports, backend_flags, outdir=orig_dir / "dist")
+    os.chdir(orig_dir)
+    return built_wheel
 
 
 # simple 'pyodide build' command
@@ -151,20 +127,47 @@ def main(
         "requested",
         help="Which symbols should be exported when linking .so files?",
     ),
+    build_dependencies: bool = typer.Option(
+        False, help="Fetch non-pyodide dependencies from pypi and build them too."
+    ),
+    skip_dependency: list[str] = typer.Option(
+        [],
+        help="Skip building or resolving a single dependency. Use multiple times or provide a comma separated list to skip multiple dependencies.",
+    ),
     ctx: typer.Context = typer.Context,
 ) -> None:
     """Use pypa/build to build a Python package from source, pypi or url."""
+    extras: list[str] = []
+    if source_location is not None:
+        extras = re.findall(r"\[(\w+)\]", source_location)
+        if len(extras) != 0:
+            source_location = source_location[0 : source_location.find("[")]
+
     if not source_location:
         # build the current folder
-        source(".", exports, ctx)
+        wheel = source(".", exports, ctx)
     elif source_location.find("://") != -1:
-        url(source_location, exports, ctx)
+        wheel = url(source_location, exports, ctx)
     elif Path(source_location).is_dir():
         # a folder, build it
-        source(source_location, exports, ctx)
+        wheel = source(source_location, exports, ctx)
+    elif source_location.find("/") == -1:
+        # try fetch or build from pypi
+        wheel = pypi(source_location, exports, ctx)
     else:
-        # try fetch from pypi
-        pypi(source_location, exports, ctx)
+        raise RuntimeError(f"Couldn't determine source type for {source_location}")
+    # now build deps for wheel
+    if build_dependencies:
+        try:
+            build_dependencies_for_wheel(
+                wheel, extras, skip_dependency, exports, ctx.args
+            )
+        except BaseException as e:
+            import traceback
+
+            print("Failed building dependencies for wheel:", traceback.format_exc())
+            wheel.unlink()
+            raise e
 
 
 main.typer_kwargs = {  # type: ignore[attr-defined]