+++ pyodide-build/pyodide_build/out_of_tree/build.py
@@ -1,10 +1,13 @@
 import os
 from pathlib import Path
+from typing import Any
 
 from .. import common, pypabuild, pywasmcross
 
 
-def run(exports, args):
+def run(exports: Any, args: list[str], outdir: Path | None = None) -> Path:
+    if outdir is None:
+        outdir = Path("./dist")
     cflags = common.get_make_flag("SIDE_MODULE_CFLAGS")
     cflags += f" {os.environ.get('CFLAGS', '')}"
     cxxflags = common.get_make_flag("SIDE_MODULE_CXXFLAGS")
@@ -12,8 +15,6 @@ def run(exports, args):
     ldflags = common.get_make_flag("SIDE_MODULE_LDFLAGS")
     ldflags += f" {os.environ.get('LDFLAGS', '')}"
 
-    curdir = Path.cwd()
-    (curdir / "dist").mkdir(exist_ok=True)
     build_env_ctx = pywasmcross.get_build_env(
         env=os.environ.copy(),
         pkgname="",
@@ -25,4 +26,5 @@ def run(exports, args):
     )
 
     with build_env_ctx as env:
-        pypabuild.build(env, " ".join(args))
+        built_wheel = pypabuild.build(env, " ".join(args), outdir=str(outdir))
+    return Path(built_wheel)