+++ docs/project/changelog.md
@@ -27,6 +27,10 @@ _January 3, 2023_
 
 [See the release notes for a summary.](https://blog.pyodide.org/posts/0.22-release/)
 
+- {{ Enhancement}} Add `--build-dependencies` to pyodide build command to fetch and build dependencies of a package being
+  built. Also adds `--skip-dependency` to ignore selected dependencies.
+  {pr}`3310`
+
 ### Deployment and testing
 
 - {{ Breaking }} `pyodide-cdn2.iodide.io` is not available anymore. Please use